refactor: interfaces, files names and files path (#111)
