+++ client/src/Hooks/index.ts
@@ -1,3 +1,3 @@
 export { default as useSnackBar, SnackBarContext } from './snackbar-hook';
 export { default as useAuth } from './use-auth';
-export { default as getQuizzes } from './get-quizzes';
+export { default as useQuizzes } from './use-quizzes';
