+++ client/src/Pages/Teacher/index.ts
@@ -1,3 +1,3 @@
 // eslint-disable-next-line import/prefer-default-export
-export { default as MyQuizzes } from './MyQuizzes';
+export { default as MyQuizzes } from './Quizzes';
 export { default as CreateQuiz } from './CreateQuiz';
